--- kotlin-build-plugins-1/settings.gradle.kts
@@ -1,5 +1,0 @@
-plugins {
-    id("org.gradle.toolchains.foojay-resolver-convention") version "0.5.0"
-}
-rootProject.name = "kotlin-build-plugins-1"
-
